+++ Fay/main.cpp
@@ -1,5 +1,5 @@
+#define MAP_EDITOR 1
 #include <Core/Core.h>
-#define MAP_EDITOR 0
 int main()
 {
 #if !MAP_EDITOR
@@ -73,21 +73,35 @@ Fay::Shader* s = new Fay::Shader("Res/Shaders/basic.vert", "Res/Shaders/basic.fr
 Fay::Shader& shader = *s;
 Fay::TileLayer* layer = new Fay::TileLayer(&shader);
 
-Fay::MapEditor editor(layer, 32, 18, 32.0f);
+Fay::Shader* s2 = new Fay::Shader("Res/Shaders/basic.vert", "Res/Shaders/basic.frag");
+Fay::Shader& shader2 = *s2;
+Fay::TileLayer* layer2 = new Fay::TileLayer(&shader2);
+
+Fay::MapEditor editor(layer, layer2, 32, 18, 32.0f);
 Fay::Camera camera(920, 540);
+Fay::Camera camera2(920, 540);
+
 
 
 shader.enable();
 shader.setUniform2f("light_pos", Fay::Vec2(4.0f, 1.5f)); // lighting
 
+shader2.enable();
+shader2.setUniform2f("light_pos", Fay::Vec2(4.0f, 1.5f)); // lighting
+
 GLint texIds[]
 {
-	0, 1, 2, 3, 4, 5, 6, 7, 9, 9
+	0, 1, 2, 3, 4, 5, 6, 7, 8, 9
 };
 shader.enable();
 shader.setUniform1iv("textures", texIds, 10);
 shader.setUniformMat4("pr_matrix", Fay::Mat4::orthographic(-16.0f, 16.0f, -9.0f, 9.0f, -1.0f, 1.0f)); // change to use camera->getProjection();
 shader.setUniformMat4("vw_matrix", camera.getViewMatrix());
+
+shader2.enable();
+shader2.setUniform1iv("textures", texIds, 10);
+shader2.setUniformMat4("pr_matrix", Fay::Mat4::orthographic(-16.0f, 16.0f, -9.0f, 9.0f, -1.0f, 1.0f)); // change to use camera->getProjection();
+shader2.setUniformMat4("vw_matrix", camera2.getViewMatrix());
 Fay::Vec3 camPos;
 
 IMGUI_CHECKVERSION();
@@ -109,6 +123,15 @@ ImGui_ImplOpenGL3_Init("#version 330");
 
 while (!window.closed())
 {
+
+	camera.setPosition(Fay::Vec3(16.0f, 9.0f, 0.0f));
+
+	shader.enable();
+	shader.setUniformMat4("vw_matrix", camera.getViewMatrix());
+
+	shader2.enable();
+	shader2.setUniformMat4("vw_matrix", camera.getViewMatrix());
+
 	window.clear();
 	ImGui_ImplOpenGL3_NewFrame();
 	ImGui_ImplGlfw_NewFrame();
@@ -118,6 +141,7 @@ while (!window.closed())
 	editor.handleInput();
 	editor.render();
 	layer->render();
+	layer2->render();
 	editor.update();
 
 	ImGui::Render();
@@ -130,13 +154,11 @@ while (!window.closed())
 		ImGui::RenderPlatformWindowsDefault();
 		glfwMakeContextCurrent(backup_current_context);
 	}
+	
 	ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
-
-	camera.setPosition(Fay::Vec3(16.0f, 9.0f, 0.0f));
-	shader.setUniformMat4("vw_matrix", camera.getViewMatrix());
-
 	window.update();
 }
+
 #endif
 	return 0;
 }